Allow users to pass env var names as flags
This commit allows users to specify an environment
variable as a flag. The codecov uploader will
use the contents of the environment variable as
the flags when it runs.

This commit checks if the string specified in
PARAM_FLAGS corresponds to an env var and if
it does it passes the contents to the flags option
of the uploader. If it doesn't then it uses the
PARAM_FLAGS string itself.

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>